Implement reviewer suggestions
- add TODO for removal of hardcoded source mapping
- improve reducer when sorting sources
- fix typo in example input file
- move source mapping to shared-constants
- fix comparison with 0
- add missing description in json schema
