+++ src/ElectronBackend/input/OpossumInputFileSchema.json
@@ -207,7 +207,7 @@
     },
     "externalAttributionSources": {
       "type": "object",
-      "description": "",
+      "description": "A mapping of attribution sources to prettified names and priority, which is used for ordering in the PackagePanel. The lower the number the further up the source appears in the PackagePanel.",
       "additionalProperties": {
         "type": "object",
         "properties": {
